MIR: Split code into several files
